Reimplement !listbans
The output is as bad as it's always been, should fix it in the future.

+++ src/game/g_utilities.cpp
@@ -155,7 +155,7 @@ string ValueForKey(int clientNum, const std::string &key) {
   return value;
 }
 
-string TimeStampToString(int t) {
+string TimeStampToString(int64_t t) {
   char buf[MAX_TOKEN_CHARS];
   struct tm *lt = NULL;
   time_t toConvert = t;
